pyproject.toml
[tool.poetry]
name = "tutoroffice"
version = "0.1.0"
description = ""
authors = ["Takeshi <orwellj@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.1.6"
djangorestframework = "^3.14.0"
python-dotenv = "^0.21.1"
pillow = "^9.4.0"
celery = {extras = ["redis"], version = "^5.2.7"}
django-filter = "^23.1"
eventlet = "^0.33.3"
drf-yasg = "^1.21.5"
psycopg2-binary = "^2.9.6"
psycopg2 = "^2.9.6"
djangorestframework-simplejwt = "^5.2.2"
black = "^23.3.0"
flake8 = "^6.0.0"
isort = "^5.12.0"


[tool.poetry.group.dev.dependencies]
django-debug-toolbar = "^4.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 79
target-version = ['py311']
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''
